moved to cassandra subpackage
