+++ cassandra-ns/src/main/resources/META-INF/services/com.excelian.mache.builder.StorageProvisioner
@@ -1,1 +1,1 @@
-com.excelian.mache.builder.CassandraProvisioner
+com.excelian.mache.cassandra.builder.CassandraProvisioner
