+++ cassandra-ns/src/test/java/com/excelian/mache/cassandra/CassandraConnectorTest.java
@@ -1,10 +1,11 @@
-package com.excelian.mache;
+package com.excelian.mache.cassandra;
 
 import com.codeaffine.test.ConditionalIgnoreRule;
 import com.codeaffine.test.ConditionalIgnoreRule.IgnoreIf;
 import com.datastax.driver.core.Cluster;
 import com.datastax.driver.core.Metadata;
 import com.datastax.driver.core.Session;
+import com.excelian.mache.NoRunningCassandraDbForTests;
 import org.junit.Rule;
 import org.junit.Test;
 import org.slf4j.Logger;
