+++ cassandra-ns/src/test/java/com/excelian/mache/cassandra/CassandraLoaderTest.java
@@ -1,8 +1,11 @@
-package com.excelian.mache;
+package com.excelian.mache.cassandra;
 
 import com.codeaffine.test.ConditionalIgnoreRule;
 import com.codeaffine.test.ConditionalIgnoreRule.IgnoreIf;
 import com.datastax.driver.core.Cluster;
+import com.excelian.mache.ExCacheLoader;
+import com.excelian.mache.NoRunningCassandraDbForTests;
+import com.excelian.mache.SchemaOptions;
 import org.junit.Rule;
 
 /**
