+++ cassandra-ns/src/test/java/com/excelian/mache/cassandra/TestCacheLoaderBase.java
@@ -1,5 +1,7 @@
-package com.excelian.mache;
+package com.excelian.mache.cassandra;
 
+import com.excelian.mache.CacheThing;
+import com.excelian.mache.ExCacheLoader;
 import com.google.common.cache.CacheLoader;
 import org.junit.After;
 import org.junit.Before;
