+++ jmeter/src/main/java/com/excelian/mache/jmeter/cassandra/WriteToCassandra.java
@@ -1,7 +1,7 @@
 package com.excelian.mache.jmeter.cassandra;
 
 import com.datastax.driver.core.Cluster;
-import com.excelian.mache.CassandraCacheLoader;
+import com.excelian.mache.cassandra.CassandraCacheLoader;
 import com.excelian.mache.SchemaOptions;
 import com.excelian.mache.jmeter.MacheAbstractJavaSamplerClient;
 import org.apache.jmeter.config.Arguments;
